requirements.txt
google-api-python-client
pandas
numpy
matplotlib
nltk
streamlit
python-dotenv
langchain
chromadb
faiss-cpu
langchain_google_genai
langchain_core
langchain_community

-e.